+++ src/test/java/roomescape/controller/api/ReservationControllerTest.java
@@ -1,10 +1,11 @@
-package roomescape.controller;
+package roomescape.controller.api;
 
 import static org.hamcrest.Matchers.is;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import java.util.Map;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -16,7 +17,7 @@ import org.springframework.test.context.jdbc.Sql;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Sql(scripts = {"/test.sql"}, executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
-class ReservationTimeControllerTest {
+class ReservationControllerTest {
 
     @LocalServerPort
     int port;
@@ -26,70 +27,67 @@ class ReservationTimeControllerTest {
         RestAssured.port = port;
     }
 
-    @DisplayName("예약 시간 목록 조회")
+    @DisplayName("존재하지 않는 예약 삭제")
     @Test
-    void getReservationTimesWhenEmpty() {
+    void deletedReservationNotFound() {
         RestAssured.given().log().all()
-                .when().get("/times")
+                .when().delete("/reservations/100")
                 .then().log().all()
-                .statusCode(200)
-                .body("size()", is(7));
+                .statusCode(400);
     }
 
-    @DisplayName("시간 목록 예약 여부 조회")
+    @DisplayName("예약 목록 조회")
     @Test
-    void getReservationTimesWithBooked() {
+    void getReservationsWhenEmpty() {
         RestAssured.given().log().all()
-                .when().get("/times?date=2024-04-29&themeId=1")
+                .when().get("/reservations")
                 .then().log().all()
                 .statusCode(200)
-                .body("size()", is(7));
+                .body("size()", is(13));
     }
 
-    @DisplayName("예약 시간 추가 및 삭제")
+    @DisplayName("예약 추가 및 삭제")
     @Test
-    void saveAndDeleteReservationTime() {
-        final Map<String, String> params = Map.of("startAt", "10:10");
+    void saveAndDeleteReservation() {
+        final Map<String, Object> params = Map.of(
+                "name", "브라운",
+                "date", "2025-08-05",
+                "timeId", 1L,
+                "themeId", 1L);
 
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
-                .when().post("/times")
+                .when().post("/reservations")
                 .then().log().all()
                 .statusCode(201)
-                .header("Location", "/times/8");
+                .header("Location", "/reservations/14");
 
         RestAssured.given().log().all()
-                .when().delete("/times/8")
+                .when().delete("/reservations/1")
                 .then().log().all()
                 .statusCode(204);
     }
 
-    @DisplayName("중복된 시간 추가")
+    @DisplayName("존재하지 않는 시간으로 예약 추가")
     @Test
-    void saveDuplicatedTime() {
-        final Map<String, String> params = Map.of("startAt", "10:00");
+    void reservationTimeForSaveNotFound() {
+        final Map<String, Object> params = Map.of(
+                "name", "브라운",
+                "date", "2025-08-05",
+                "timeId", 100L);
 
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
-                .when().post("/times")
+                .when().post("/reservations")
                 .then().log().all()
                 .statusCode(400);
     }
 
-    @DisplayName("존재하지 않는 예약 시간 삭제")
-    @Test
-    void deleteReservationTimeNotFound() {
-        RestAssured.given().log().all()
-                .when().delete("/times/100")
-                .then().log().all()
-                .statusCode(400);
-    }
-
-    @DisplayName("유효하지 않은 시간 형식 입력")
+    @DisplayName("유효하지 않은 날짜 형식 입력")
     @ParameterizedTest
-    @ValueSource(strings = {"", "    ", "11:11:11", "25:10"})
+    @ValueSource(strings = {"", "    ", "2022.22.11"})
     void invalidTimeFormat(final String time) {
         final Map<String, String> params = Map.of("startAt", time);
