+++ src/test/java/roomescape/controller/render/AdminControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.controller.render;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.BeforeEach;
